Parse `type_builder_block` in `value_expression_block`
